applied changes from pull request review
